requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.4
openai-whisper==20240930
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
webrtcvad==2.0.10
